+++ src/julia-syntax.scm
@@ -1807,10 +1807,10 @@
                     (make-fuse f x)))
               (let ((f (cadr e)))
                 (cond ((dotop-named? f)
-                       (make-fuse- (undotop f) (cddr e)))
+                       (make-fuse- (unshortcircuit (undotop f)) (cddr e)))
                       ;; (.+)(a, b) is parsed as (call (|.| +) a b), but we still want it to fuse
                       ((and (length= f 2) (eq? (car f) '|.|))
-                       (make-fuse- (cadr f) (cddr e)))
+                       (make-fuse- (unshortcircuit (cadr f)) (cddr e)))
                       (else
                         e))))
             e)))
@@ -2055,6 +2055,12 @@
             (cons (car e)
                   (map expand-forms (cdr e)))))))
 
+(define (unshortcircuit op)
+  (case op
+    ((&&) `(top andand))
+    ((|\|\||) `(top oror))
+    (else op)))
+
 ;; table mapping expression head to a function expanding that form
 (define expand-table
   (table
@@ -2109,9 +2115,14 @@
    (lambda (e)
      (if (length= e 2)
          ;; e = (|.| op)
-         `(call (top BroadcastFunction) ,(cadr e))
+         `(call (top BroadcastFunction) ,(unshortcircuit (cadr e)))
          ;; e = (|.| f x)
          (expand-fuse-broadcast '() e)))
+
+   '.&&
+   (lambda (e) (expand-fuse-broadcast '() `(|.| ,(unshortcircuit (undotop (car e))) (tuple ,@(cdr e)))))
+   '|.\|\||
+   (lambda (e) (expand-fuse-broadcast '() `(|.| ,(unshortcircuit (undotop (car e))) (tuple ,@(cdr e)))))
 
    '.=
    (lambda (e)
@@ -2293,10 +2304,10 @@
      (if (length> e 2)
          (let ((f (cadr e)))
            (cond ((dotop-named? f)
-                  (expand-fuse-broadcast '() `(|.| ,(undotop f) (tuple ,@(cddr e)))))
+                  (expand-fuse-broadcast '() `(|.| ,(unshortcircuit (undotop f)) (tuple ,@(cddr e)))))
                  ;; "(.op)(...)"
                  ((and (length= f 2) (eq? (car f) '|.|))
-                  (expand-fuse-broadcast '() `(|.| ,(cadr f) (tuple ,@(cddr e)))))
+                  (expand-fuse-broadcast '() `(|.| ,(unshortcircuit (cadr f)) (tuple ,@(cddr e)))))
                  ((eq? f 'ccall)
                   (if (not (length> e 4)) (error "too few arguments to ccall"))
                   (let* ((cconv (cadddr e))